data consistency

+++ app/Http/Controllers/UserManagementController.php
@@ -5,7 +5,9 @@ namespace App\Http\Controllers;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 use Illuminate\View\View;
-
+use App\Models\UserLogin;
+use App\Models\User;
+use App\Models\Department;
 
 class UserManagementController extends Controller
 {
@@ -14,10 +16,14 @@ class UserManagementController extends Controller
         return view('user-management');
     }
 
-    public function destroy($id): View
+    public function destroy($id)
     {
+        $user = User::find($id);
+        if ($match = Department::find($user->department_id)){
+            return redirect('/user-management')->with('errorBossDepartment', "このユーザーは「{$match->department_name}」の責任者のため削除できません。");
+        }
         $user = UserLogin::find($id);
         $user->delete();
-        return redirect('/user-management');
+        return redirect('/user-management')->with('successDeleteUser', 'ユーザーを削除しました。');
     }
 }
